face_recognition/requirements.txt
gradio>=6.0.0
facenet-pytorch>=2.6.0
httpx>=0.27.0
numpy>=2.0.0
opencv-python>=4.9.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.2.0
